Mentions in source suggestions trigger notifications
These previously weren't working because comments on source suggestion
submissions aren't "real" comments, just a text field on the actual
SourceSuggestion entity. Ideally we'd refactor it so that they were
real Comment entities, but that's slightly more involved. For now, this
makes it so that the NotifyMentioned background job can accept arbitrary
text and an Event instead of a Comment and an Event, which works in both
cases.