Include admins when editors are mentioned.
More specifically, now when you mention a role by name, members of that
role and all roles higher than it will be notified. In practice right
now that means when editors are mentioned, admins will be too. You
cannot (and never have been able to) @mention curators as that's
everyone on the site.