Remove rescheduling from ensure blocks.
This will prevent runaway job failures like we saw with the sequence
ontology update task. after_perform is only called if the job succeeds,
so we will no longer retry in delayed_job AND reschedule a duplicate
copy. Also reduce the number of retries from the default 25 to a more
sane 3.