update ndex script according to rudi's instructions
"All nodes should have their “Node Type” attribute renamed to type and its value set either to gene or variant or disease.
As far as I can see, the values are already in good shape, so only the attribute name needs to be modified.
The changes should not require more than 5 minutes and need to be made in the section of the script that bears the “4 LOAD PLANS’"