Handle coordinates passed in with commas
This adds a new mixin that can be generically used anywhere we're
getting a string from the frontend that we eventually cast to a number.

When mixed in, it defines a class method that allows you to specify
which columns require this treatment. The specified columns will have
their default setter method overidden to remove commas before the
ActiveRecord column casting logic.

closes griffithlab/civic-client#1315