Use the nightly civicpy cache pkl location
Currently, CIViCpy would look for the cache pkl file in the default
location ~/.civicpy/cache.pkl. If that file doesn't exit or is
out-of-date, CIViCpy will download the nightly civicpy cache pkl file to
this default location and then use it to create the VCFs.

This is silly when the nightly pkl file lives on
the same machine so this PR sets the CIVICPY_CACHE_FILE environment
variable to point to the nightly pkl file. This will overwrite the default
and result in the nightly pkl file being used directly to create the
VCFs.